Organise generator into namespaces
Trying to make this project as easy to contribute to as possible - smaller modules make it easier to focus on parts relevant to an issue
