+++ generator/safe.php
@@ -4,9 +4,9 @@
 
 require __DIR__.'/vendor/autoload.php';
 
-use Safe\GenerateCommand;
-use Safe\ScanObjectsCommand;
-use Safe\FunctionInfoCommand;
+use Safe\Commands\FunctionInfoCommand;
+use Safe\Commands\GenerateCommand;
+use Safe\Commands\ScanObjectsCommand;
 use Symfony\Component\Console\Application;
 
 $application = new Application();
