+++ generator/src/Commands/GenerateCommand.php
@@ -2,7 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Safe;
+namespace Safe\Commands;
+
+use Safe\XmlDocParser\Scanner;
+use Safe\XmlDocParser\DocPage;
+use Safe\Generator\FileCreator;
+use Safe\Generator\ComposerJsonEditor;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
