+++ generator/src/Commands/ScanObjectsCommand.php
@@ -2,8 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Safe;
+namespace Safe\Commands;
 
+use Safe\XmlDocParser\Scanner;
+use Safe\XmlDocParser\DocPage;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
