+++ generator/tests/PhpStanFunctions/PhpStanTypeTest.php
@@ -4,13 +4,12 @@ declare(strict_types=1);
 
 namespace Safe\PhpStanFunctions;
 
-
 use PHPUnit\Framework\TestCase;
-use Safe\Method;
+use Safe\XmlDocParser\Method;
 
 class PhpStanTypeTest extends TestCase
 {
-    public function testMixedTypes(): void 
+    public function testMixedTypes(): void
     {
         $param = new PhpStanType('array|string|int');
         $this->assertEquals('array|string|int', $param->getDocBlockType());
@@ -54,8 +53,8 @@ class PhpStanTypeTest extends TestCase
         $this->assertEquals('array{0:float,1:float,2:float,3:float,4:float,5:float}', $param->getDocBlockType());
         $this->assertEquals('array', $param->getSignatureType());
     }
-    
-    public function testNullable(): void 
+
+    public function testNullable(): void
     {
         $param = new PhpStanType('array|null');
         $this->assertEquals(true, $param->isNullable());
@@ -77,8 +76,8 @@ class PhpStanTypeTest extends TestCase
         $this->assertEquals('\HashContext|null', $param->getDocBlockType());
         $this->assertEquals('?\HashContext', $param->getSignatureType());
     }
-    
-    public function testParenthesisOutsideOfCallable(): void 
+
+    public function testParenthesisOutsideOfCallable(): void
     {
         $param = new PhpStanType('(?int)|(?string)');
         $this->assertEquals(true, $param->isNullable());
@@ -93,8 +92,8 @@ class PhpStanTypeTest extends TestCase
         $this->assertEquals('string|false', $param->getDocBlockType());
         $this->assertEquals('string', $param->getSignatureType());
     }
-    
-    public function testResource(): void 
+
+    public function testResource(): void
     {
         $param = new PhpStanType('resource');
         $this->assertEquals('resource', $param->getDocBlockType());
@@ -107,8 +106,8 @@ class PhpStanTypeTest extends TestCase
         $this->assertEquals('\GMP', $param->getDocBlockType());
         $this->assertEquals('\GMP', $param->getSignatureType());
     }
-    
-    public function testVoid(): void 
+
+    public function testVoid(): void
     {
         $param = new PhpStanType('');
         $this->assertEquals('', $param->getDocBlockType());
@@ -122,8 +121,8 @@ class PhpStanTypeTest extends TestCase
         $this->assertEquals('void', $param->getDocBlockType());
         $this->assertEquals('void', $param->getSignatureType());
     }
-    
-    public function testOciSpecialCases(): void 
+
+    public function testOciSpecialCases(): void
     {
         $param = new PhpStanType('OCI-Collection');
         $this->assertEquals('\OCI-Collection', $param->getDocBlockType());
@@ -133,14 +132,14 @@ class PhpStanTypeTest extends TestCase
         $this->assertEquals('\OCI-Lob', $param->getDocBlockType());
         $this->assertEquals('', $param->getSignatureType());
     }
-    
-    public function testErrorTypeInteraction(): void 
+
+    public function testErrorTypeInteraction(): void
     {
         //bool => void if the method is falsy
         $param = new PhpStanType('bool');
         $this->assertEquals('void', $param->getDocBlockType(Method::FALSY_TYPE));
         $this->assertEquals('void', $param->getSignatureType(Method::FALSY_TYPE));
-        
+
         //int|false => int if the method is falsy
         $param = new PhpStanType('int|false');
         $this->assertEquals('int', $param->getDocBlockType(Method::FALSY_TYPE));
@@ -155,8 +154,8 @@ class PhpStanTypeTest extends TestCase
         $this->assertEquals('array|null', $param->getDocBlockType(Method::FALSY_TYPE));
         $this->assertEquals('?array', $param->getSignatureType(Method::FALSY_TYPE));
     }
-    
-    public function testDuplicateType(): void 
+
+    public function testDuplicateType(): void
     {
         $param = new PhpStanType('array<string,string>|array<string,false>|array<string,array<int,mixed>>');
         $this->assertEquals('array<string,string>|array<string,false>|array<string,array<int,mixed>>', $param->getDocBlockType());
